+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 2.2.1 - unreleased
+## 2.2.1 - 2019-09-05
 
 ### Fixed
 
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 -   autosuggestField support for environment settings
+-   customer info sent
+-   shortNumber used for orderId
 
 ## 2.2.0 - 2019-09-03